\name{lessOptimizersTemplateR-package}
\alias{lessOptimizersTemplateR-package}
\alias{lessOptimizersTemplateR}
\docType{package}
\title{
\packageTitle{lessOptimizersTemplateR}
}
\description{
\packageDescription{lessOptimizersTemplateR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lessOptimizersTemplateR}
\packageIndices{lessOptimizersTemplateR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lessOptimizersTemplateR}

Maintainer: \packageMaintainer{lessOptimizersTemplateR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
